digital-forensics/lab-2/requirements.txt
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
fonttools==4.54.1
imageio==2.35.1
kiwisolver==1.4.7
lazy_loader==0.4
matplotlib==3.9.2
mypy-extensions==1.0.0
networkx==3.3
numpy==2.1.1
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
pyparsing==3.1.4
python-dateutil==2.9.0.post0
PyWavelets==1.7.0
scikit-image==0.24.0
scipy==1.14.1
six==1.16.0
tifffile==2024.9.20
typing_extensions==4.12.2
